Move Shield::Model to "shield/model"
